@@ -1,7 +1,4 @@
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_colorpicker/flutter_colorpicker.dart';
 
 @immutable
@@ -12,7 +9,10 @@ class ColorSelector extends StatelessWidget {
   final double width;
 
   const ColorSelector(
-      {this.height = 40, this.width = 40, this.color, this.onColorChanged});
+      {this.height = 40,
+      this.width = 40,
+      required this.color,
+      required this.onColorChanged});
 
   @override
   Widget build(BuildContext context) {
@@ -49,7 +49,7 @@ class ColorSelector extends StatelessWidget {
               ),
             ),
             actions: <Widget>[
-              FlatButton(
+              TextButton(
                 child: const Text('Close'),
                 onPressed: () {
                   Navigator.of(context).pop();
