@@ -28,7 +28,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatelessWidget {
-  MyHomePage({Key key}) : super(key: key);
+  MyHomePage({super.key});
 
   Widget build(BuildContext context) {
     return Scaffold(
@@ -58,7 +58,7 @@ class MyHomePage extends StatelessWidget {
             NeumorphicButton(
                 margin: EdgeInsets.only(top: 12),
                 onPressed: () {
-                  NeumorphicTheme.of(context).themeMode =
+                  NeumorphicTheme.of(context)?.themeMode =
                       NeumorphicTheme.isUsingDark(context)
                           ? ThemeMode.light
                           : ThemeMode.dark;
@@ -98,10 +98,10 @@ class MyHomePage extends StatelessWidget {
     );
   }
 
-  Color _iconsColor(BuildContext context) {
+  Color? _iconsColor(BuildContext context) {
     final theme = NeumorphicTheme.of(context);
-    if (theme.isUsingDark) {
-      return theme.current.accentColor;
+    if (theme?.isUsingDark ?? false) {
+      return theme?.current?.accentColor;
     } else {
       return null;
     }
