@@ -1,5 +1,4 @@
 import 'package:example/lib/top_bar.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class ClockAlarmPage extends StatelessWidget {
@@ -230,8 +229,8 @@ class Alarm {
   final String label;
 
   const Alarm({
-    @required this.enabled,
-    @required this.time,
-    @required this.label,
+    required this.enabled,
+    required this.time,
+    required this.label,
   });
 }
