@@ -1,6 +1,5 @@
 import 'package:example/lib/ThemeConfigurator.dart';
 import 'package:example/lib/top_bar.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class FormSample extends StatelessWidget {
@@ -35,7 +34,7 @@ class __PageState extends State<_Page> {
   String firstName = "";
   String lastName = "";
   double age = 12;
-  Gender gender;
+  late Gender gender;
   Set<String> rides = Set();
 
   @override
@@ -181,7 +180,7 @@ class _AgeField extends StatelessWidget {
   final double age;
   final ValueChanged<double> onChanged;
 
-  _AgeField({@required this.age, this.onChanged});
+  _AgeField({required this.age, required this.onChanged});
 
   @override
   Widget build(BuildContext context) {
@@ -230,14 +229,15 @@ class _TextField extends StatefulWidget {
 
   final ValueChanged<String> onChanged;
 
-  _TextField({@required this.label, @required this.hint, this.onChanged});
+  _TextField(
+      {required this.label, required this.hint, required this.onChanged});
 
   @override
   __TextFieldState createState() => __TextFieldState();
 }
 
 class __TextFieldState extends State<_TextField> {
-  TextEditingController _controller;
+  late TextEditingController _controller;
 
   @override
   void initState() {
@@ -283,8 +283,8 @@ class _GenderField extends StatelessWidget {
   final ValueChanged<Gender> onChanged;
 
   const _GenderField({
-    @required this.gender,
-    @required this.onChanged,
+    required this.gender,
+    required this.onChanged,
   });
 
   @override
@@ -313,7 +313,9 @@ class _GenderField extends StatelessWidget {
               ),
               value: Gender.MALE,
               child: Icon(Icons.account_box),
-              onChanged: (value) => this.onChanged(value),
+              onChanged: (value) {
+                if (value != null) this.onChanged(value);
+              },
             ),
             SizedBox(width: 12),
             NeumorphicRadio(
@@ -324,7 +326,9 @@ class _GenderField extends StatelessWidget {
               ),
               value: Gender.FEMALE,
               child: Icon(Icons.pregnant_woman),
-              onChanged: (value) => this.onChanged(value),
+              onChanged: (value) {
+                if (value != null) this.onChanged(value);
+              },
             ),
             SizedBox(width: 12),
             NeumorphicRadio(
@@ -335,7 +339,9 @@ class _GenderField extends StatelessWidget {
               ),
               value: Gender.NON_BINARY,
               child: Icon(Icons.supervised_user_circle),
-              onChanged: (value) => this.onChanged(value),
+              onChanged: (value) {
+                if (value != null) this.onChanged(value);
+              },
             ),
             SizedBox(
               width: 18,
