@@ -1,9 +1,7 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class AppBarWidgetPage extends StatelessWidget {
-  AppBarWidgetPage({Key key}) : super(key: key);
+  AppBarWidgetPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -152,7 +150,7 @@ class FirstThemeContent extends StatelessWidget {
 class _MyDrawer extends StatelessWidget {
   final bool isLead;
 
-  const _MyDrawer({Key key, this.isLead = true}) : super(key: key);
+  const _MyDrawer({super.key, this.isLead = true});
 
   @override
   Widget build(BuildContext context) {
