@@ -1,11 +1,10 @@
 import 'package:example/lib/Code.dart';
 import 'package:example/lib/ThemeConfigurator.dart';
 import 'package:example/lib/top_bar.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class BackgroundWidgetPage extends StatefulWidget {
-  BackgroundWidgetPage({Key key}) : super(key: key);
+  BackgroundWidgetPage({super.key});
 
   @override
   createState() => _WidgetPageState();
