@@ -1,10 +1,9 @@
 import 'package:example/lib/ThemeConfigurator.dart';
 import 'package:example/lib/top_bar.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class IconWidgetPage extends StatefulWidget {
-  IconWidgetPage({Key key}) : super(key: key);
+  IconWidgetPage({super.key});
 
   @override
   createState() => _WidgetPageState();
@@ -741,7 +740,7 @@ class _PageState extends State<_Page> {
     Icons.picture_in_picture,
     Icons.picture_in_picture_alt,
     Icons.pie_chart,
-    Icons.pie_chart_outlined,
+    Icons.pie_chart_outline,
     Icons.pin_drop,
     Icons.place,
     Icons.play_arrow,
