@@ -2,13 +2,12 @@ import 'package:example/lib/Code.dart';
 import 'package:example/lib/ThemeConfigurator.dart';
 import 'package:example/lib/color_selector.dart';
 import 'package:example/lib/top_bar.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 import 'dart:math' show Random;
 
 class IndicatorWidgetPage extends StatefulWidget {
-  IndicatorWidgetPage({Key key}) : super(key: key);
+  IndicatorWidgetPage({super.key});
 
   @override
   createState() => _WidgetPageState();
@@ -102,7 +101,7 @@ NeumorphicIndicator(
             percent: percent,
           ),
           SizedBox(width: 12),
-          FlatButton(
+          TextButton(
               child: Text('Update'),
               onPressed: () {
                 setState(() {
@@ -298,7 +297,7 @@ NeumorphicIndicator(
             duration: Duration(seconds: 1),
           ),
           SizedBox(width: 12),
-          FlatButton(
+          TextButton(
               child: Text('Update'),
               onPressed: () {
                 setState(() {
@@ -356,7 +355,7 @@ NeumorphicIndicator(
               percent: percent,
               curve: Curves.bounceOut),
           SizedBox(width: 12),
-          FlatButton(
+          TextButton(
               child: Text('Update'),
               onPressed: () {
                 setState(() {
