@@ -1,11 +1,10 @@
 import 'package:example/lib/Code.dart';
 import 'package:example/lib/ThemeConfigurator.dart';
 import 'package:example/lib/top_bar.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class RadioButtonWidgetPage extends StatefulWidget {
-  RadioButtonWidgetPage({Key key}) : super(key: key);
+  RadioButtonWidgetPage({super.key});
 
   @override
   createState() => _WidgetPageState();
@@ -69,7 +68,7 @@ class _DefaultWidget extends StatefulWidget {
 }
 
 class _DefaultWidgetState extends State<_DefaultWidget> {
-  int groupValue;
+  late int groupValue;
 
   Widget _buildCode(BuildContext context) {
     return Code("""
@@ -108,7 +107,7 @@ NeumorphicRadio(
             value: 1991,
             onChanged: (value) {
               setState(() {
-                groupValue = value;
+                if (value != null) groupValue = value;
               });
             },
             padding: EdgeInsets.all(8.0),
@@ -120,7 +119,7 @@ NeumorphicRadio(
             groupValue: groupValue,
             onChanged: (value) {
               setState(() {
-                groupValue = value;
+                if (value != null) groupValue = value;
               });
             },
             padding: EdgeInsets.all(8.0),
@@ -132,7 +131,7 @@ NeumorphicRadio(
             value: 2012,
             onChanged: (value) {
               setState(() {
-                groupValue = value;
+                if (value != null) groupValue = value;
               });
             },
             padding: EdgeInsets.all(8.0),
@@ -160,7 +159,7 @@ class CircleRadios extends StatefulWidget {
 }
 
 class _CircleRadiosState extends State<CircleRadios> {
-  String groupValue;
+  late String groupValue;
 
   Widget _buildCode(BuildContext context) {
     return Code("""
@@ -198,7 +197,7 @@ NeumorphicRadio(
             value: "A",
             onChanged: (value) {
               setState(() {
-                groupValue = value;
+                if (value != null) groupValue = value;
               });
             },
             padding: EdgeInsets.all(18.0),
@@ -213,7 +212,7 @@ NeumorphicRadio(
             groupValue: groupValue,
             onChanged: (value) {
               setState(() {
-                groupValue = value;
+                if (value != null) groupValue = value;
               });
             },
             padding: EdgeInsets.all(18.0),
@@ -228,7 +227,7 @@ NeumorphicRadio(
             value: "C",
             onChanged: (value) {
               setState(() {
-                groupValue = value;
+                if (value != null) groupValue = value;
               });
             },
             padding: EdgeInsets.all(18.0),
@@ -256,7 +255,7 @@ class _EnabledDisabledWidget extends StatefulWidget {
 }
 
 class _EnabledDisabledWidgetState extends State<_EnabledDisabledWidget> {
-  int groupValue;
+  late int groupValue;
 
   Widget _buildWidget(BuildContext context) {
     return Padding(
@@ -275,7 +274,7 @@ class _EnabledDisabledWidgetState extends State<_EnabledDisabledWidget> {
             child: Text("First"),
             onChanged: (value) {
               setState(() {
-                groupValue = value;
+                if (value != null) groupValue = value;
               });
             },
           ),
@@ -293,7 +292,7 @@ class _EnabledDisabledWidgetState extends State<_EnabledDisabledWidget> {
             child: Text("Second"),
             onChanged: (value) {
               setState(() {
-                groupValue = value;
+                if (value != null) groupValue = value;
               });
             },
           ),
@@ -303,7 +302,7 @@ class _EnabledDisabledWidgetState extends State<_EnabledDisabledWidget> {
   }
 
   Widget _buildCode(BuildContext context) {
-    return Code("""    
+    return Code("""
 int groupValue;
 
 NeumorphicRadio(
