@@ -2,11 +2,10 @@ import 'package:example/lib/Code.dart';
 import 'package:example/lib/ThemeConfigurator.dart';
 import 'package:example/lib/color_selector.dart';
 import 'package:example/lib/top_bar.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class SwitchWidgetPage extends StatefulWidget {
-  SwitchWidgetPage({Key key}) : super(key: key);
+  SwitchWidgetPage({super.key});
 
   @override
   createState() => _WidgetPageState();
@@ -110,7 +109,7 @@ NeumorphicSwitch(
             },
           ),
           SizedBox(width: 12),
-          FlatButton(
+          TextButton(
               onPressed: () {
                 setState(() {
                   isEnabled = !isEnabled;
