@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 import 'ThemeColorSelector.dart';
@@ -50,7 +49,7 @@ class ThemeConfigurator extends StatelessWidget {
 class _ThemeConfiguratorDialog extends StatefulWidget {
   final BuildContext contextContainingTheme;
 
-  _ThemeConfiguratorDialog({this.contextContainingTheme});
+  _ThemeConfiguratorDialog({required this.contextContainingTheme});
 
   @override
   _ThemeConfiguratorState createState() => _ThemeConfiguratorState();
@@ -82,12 +81,12 @@ class _ThemeConfiguratorState extends State<_ThemeConfiguratorDialog> {
           child: Slider(
             min: Neumorphic.MIN_INTENSITY, //in case of != 0
             max: Neumorphic.MAX_INTENSITY,
-            value: intensity,
+            value: intensity ?? 0.0,
             onChanged: (value) {
               setState(() {
                 NeumorphicTheme.update(
                   widget.contextContainingTheme,
-                  (current) => current.copyWith(
+                  (current) => current!.copyWith(
                     intensity: value,
                   ),
                 );
@@ -99,7 +98,7 @@ class _ThemeConfiguratorState extends State<_ThemeConfiguratorDialog> {
           padding: EdgeInsets.only(right: 12),
           child: Container(
             width: 40,
-            child: Text(((intensity * 100).floor() / 100).toString()),
+            child: Text(((intensity ?? 1 * 100).floor() / 100).toString()),
           ),
         ),
       ],
@@ -119,12 +118,12 @@ class _ThemeConfiguratorState extends State<_ThemeConfiguratorDialog> {
           child: Slider(
             min: Neumorphic.MIN_DEPTH,
             max: Neumorphic.MAX_DEPTH,
-            value: depth,
+            value: depth ?? 0.0,
             onChanged: (value) {
               setState(() {
                 NeumorphicTheme.update(
                   widget.contextContainingTheme,
-                  (current) => current.copyWith(depth: value),
+                  (current) => current!.copyWith(depth: value),
                 );
               });
             },
@@ -134,7 +133,7 @@ class _ThemeConfiguratorState extends State<_ThemeConfiguratorDialog> {
           padding: EdgeInsets.only(right: 12),
           child: Container(
             width: 40,
-            child: Text(depth.floor().toString()),
+            child: Text(depth?.floor().toString() ?? '0'),
           ),
         ),
       ],
