@@ -1,7 +1,6 @@
 import 'package:example/lib/top_bar.dart';
 import 'package:example/widgets/appbar/widget_app_bar.dart';
 import 'package:example/widgets/toggle/widget_toggle.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 import 'background/widget_background.dart';
@@ -18,7 +17,7 @@ import 'slider/widget_slider.dart';
 import 'switch/widget_switch.dart';
 
 class WidgetsHome extends StatelessWidget {
-  Widget _buildButton({String text, VoidCallback onClick}) {
+  Widget _buildButton({required String text, required VoidCallback onClick}) {
     return NeumorphicButton(
       margin: EdgeInsets.only(bottom: 12),
       padding: EdgeInsets.symmetric(
